fix(web-client): address CodeRabbit findings for PWA hooks
- Add error handling to promptInstall onClick in UserMenu
- Move toast logic from useOnlineStatus hook to NetworkStatusListener
  component to use i18n instead of hardcoded Ukrainian strings
- Add pwa.connectionRestored i18n key for both locales
