+++ apps/web-client/src/shared/i18n/locales/en.json
@@ -1578,6 +1578,7 @@
     "offlineDescription": "It looks like there's no internet connection. Check your connection and try again.",
     "retry": "Try again",
     "share": "Share",
-    "copied": "Copied"
+    "copied": "Copied",
+    "connectionRestored": "Connection restored"
   }
 }
