+++ apps/web-client/src/shared/i18n/locales/uk.json
@@ -1578,6 +1578,7 @@
     "offlineDescription": "Схоже, що немає з'єднання з інтернетом. Перевірте підключення та спробуйте ще раз.",
     "retry": "Спробувати знову",
     "share": "Поділитись",
-    "copied": "Скопійовано"
+    "copied": "Скопійовано",
+    "connectionRestored": "З'єднання відновлено"
   }
 }
